Odstraneni warningu + začátek komentování

+++ ppr/LISKA_PPR/PPR_Liska/DataLoader/DataLoader.h
@@ -12,18 +12,48 @@
 #define DATALOADERDLLEXPORT __declspec(dllimport)  
 #endif
 namespace DataLoader {
-
+	/// <summary>
+	/// Trida slouzi pro nacteni dat
+	/// </summary>
 	class DATALOADERDLLEXPORT DataLoader
 	{
 	public:
+		/// <summary>
+		/// 
+		/// </summary>
+		/// <param name="dbPath">Cesta k databazi, odkud se budou data nacitat</param>
 		DataLoader(char * dbPath);
-		int getSegmentIds(std::vector<int> *getSegmentIds);
+		/// <summary>
+		/// Metoda nacte id segmentu z databaze
+		/// </summary>
+		/// <param name="getSegmentIds"></param>
+		/// <returns></returns>
+		size_t getSegmentIds(std::vector<int> *getSegmentIds);
+		/// <summary>
+		/// Metoda slouzi k nacteni dat z databaze
+		/// </summary>
+		/// <param name="data">Kam se data nactou</param>
+		/// <param name="segmentId">Id segmentu, jehoz data nas zajimaji</param>
+		/// <returns>Pocet nactenych dat</returns>
 		size_t loadData(std::vector<Common::TMeasuredValue*> *data, int * segmentId);
+		/// <summary>
+		/// Rozdeli v stupni data podle jejich datumu na dny
+		/// </summary>
+		/// <param name="days">Jedna polozka vektoru = 1 den</param>
+		/// <param name="data">Data, ze kterych se cte</param>
 		void splitIntoDays(std::vector<Common::SegmentDay> *days, std::vector<Common::TMeasuredValue *> *data);
 		
 		~DataLoader();
 	private:
+		/// <summary>
+		/// Cesta k databazi
+		/// </summary>
 		const char * dbPath;
+		/// <summary>
+		/// Vraci index od do pro rozdeleni dat podle dne. Tedy par FromTo vzdy reprezentuje indexi v datech od od pro rozdeleni na den
+		/// </summary>
+		/// <param name="data"></param>
+		/// <returns>Rozsahy indexu jednotlivych dnu v datech</returns>
 		std::vector<FromTo> getDaysIndexes(std::vector<Common::TMeasuredValue*> *data);
 	};
